+++ search-films/src/components/Input/index.jsx
@@ -2,13 +2,17 @@ import cs from 'classnames';
 import { forwardRef } from 'react';
 import styles from './index.module.css';
 
+
 /**
  * Input util component (can have all props for input DOM element)
- * @param {string} className additional CSS class
- * @inputValue {string} input changeable value
- * @callback {function} onChange handler for change event
+ * @param {string} className - additional CSS class
+ * @param {string} inputValue - changeable value
+ * @param {eventHandler} onChange handler for change event
+ * @param {boolean} hasIcon (byDefault=false) true if has an icon
+ * @param {string} iconClassName additional CSS class for icon (used for setting icon image)
+ * 
  */
-const Input = forwardRef(function Input({className, inputValue, onChange, hasIcon, iconClassName, ...props}, ref) {
+const Input = forwardRef(function Input({className, inputValue, onChange, hasIcon=false, iconClassName, ...props}, ref) {
 	return (
 		<div className={cs({[styles['icon']]: /*display search icon*/ hasIcon, 
 			[iconClassName]: hasIcon})}>