+++ search-films/src/components/index.jsx
@@ -1,10 +1,10 @@
-import Flex from '../components/Flex';
-import LinkedMenuItem from '../components/LinkedMenuItem';
-import Logo from '../components/Logo';
-import Menu from '../components/Menu';
-import MovieCard from '../components/MovieCard';
-import Paragraph from '../components/Paragraph';
-import Search from '../components/Search';
-import Title from '../components/Title';
+import Flex from './Flex';
+import LinkedMenuItem from './LinkedMenuItem';
+import Logo from './Logo';
+import Menu from './Menu';
+import MovieCard from './MovieCard';
+import Paragraph from './Paragraph';
+import Search from './Search';
+import Title from './Title';
 export { Flex, LinkedMenuItem, Logo, Menu, MovieCard, Paragraph, Search, Title };