+++ search-films/src/context/user.context/index.jsx
@@ -1,0 +1,4 @@
+import UserContext from './user.context';
+import UserProvider from './user.provider';
+export { UserContext, UserProvider };
+