+++ search-films/src/context/user.context/user.context.js
@@ -1,5 +1,5 @@
 import { createContext } from 'react';
 
-const UserContext = createContext({
+const UserContext = createContext({userName: null});
 
-});
+export default UserContext;